Switch curve texture from RGBA32F to RGBA16F
Halves texture memory and bandwidth for curve control points. FP16
precision (3-4 decimal digits) is sufficient for em-space coordinates
in [0,1] range. GPU promotes to FP32 in the sampler automatically.

Setup time drops from 60ms to 2.4ms due to halved allocation size.
Rendering quality is identical (verified via screenshot comparison).

+++ src/snail.zig
@@ -67,7 +67,7 @@ pub const Atlas = struct {
     font: *const Font,
 
     // GPU texture data (CPU-side, ready for upload)
-    curve_data: []f32,
+    curve_data: []u16,
     curve_width: u32,
     curve_height: u32,
     band_data: []u16,
